fix: use u8 for Candidate.id, align proto package with directory
- Change Candidate.id from i64 to u8 to match domain semantics (candidate
  numbers are u8) — update gRPC conversion, cast_vote handler, and tests
- Change proto package from "admin" to "proto.admin" to match the
  proto/admin/ directory layout (Buf PACKAGE_DIRECTORY_MATCH)

+++ src/grpc/admin.rs
@@ -106,7 +106,7 @@ impl Admin for AdminService {
 
         // Atomically insert candidate only if election is open
         let candidate = Candidate {
-            id: req.id as i64,
+            id: req.id as u8,
             election_id: req.election_id,
             name: req.name,
         };
